changed to smartpointers + expanded parser to check for directories
