+++ src/src_normal/config/ConfigLocation.cpp
@@ -54,7 +54,6 @@ ConfigLocation::location_pointer ConfigLocation::getLocationBlock()
 
 void ConfigLocation::initLocationBlock()
 {
-	// _location_block = new LocationBlock;
 	_location_block = SmartPointer<LocationBlock>(new LocationBlock);
 	_location_block->_path = _path;
 	_location_block->_root = _root;
@@ -65,102 +64,3 @@ void ConfigLocation::initLocationBlock()
 	_location_block->_return = _return;
 	_location_block->_location_flag = _location_flag;
 }
-
-/* Debugging */
-void	ConfigLocation::print() const
-{
-	printPath();
-	printRoot();
-	printIndex();
-	printAutoIndex();
-	printAllowedMethods();
-	printClientBodySize();
-	printCgi();
-}
-
-void	ConfigLocation::printPath() const
-{
-	std::cout << "  \t" WHITE_BOLD << "Path:" RESET_COLOR " [";
-	if (!_path.empty())
-	{
-		std::cout << _path;
-	}
-	std::cout << ']' << std::endl;
-}
-
-void	ConfigLocation::printRoot() const
-{
-	std::cout << "  \t" WHITE_BOLD << "Root:" RESET_COLOR " [";
-	if (!_root.empty())
-	{	
-		std::cout << _root;
-	}
-	std::cout << ']' << std::endl;
-}
-
-void ConfigLocation::printClientBodySize() const
-{
-	std::cout << "  \t" WHITE_BOLD << "Client Body Size:" RESET_COLOR " [";
-	std::cout << _client_body_size;
-	std::cout << ']' << std::endl;
-}
-
-void ConfigLocation::printAllowedMethods() const
-{
-		std::cout << "  \t" WHITE_BOLD << "Allowed Methods:" RESET_COLOR " [";
-	for (size_t i = 0; i < _allowed_methods.size(); i++)
-	{
-		if (i != 0)
-		{
-			std::cout << ", ";
-		}
-		std::cout << _allowed_methods[i];
-	}
-	std::cout << ']' << std::endl;
-}
-
-void ConfigLocation::printAutoIndex() const
-{
-	std::cout << "  \t" WHITE_BOLD << "Autoindex:" RESET_COLOR " [";
-	if (_autoindex_status == 1)
-	{
-		std::cout << "ON";
-	}
-	else
-	{
-		std::cout << "OFF";
-	}
-	std::cout << ']' << std::endl;
-}
-
-void ConfigLocation::printIndex() const
-{
-		std::cout << "  \t" WHITE_BOLD << "Index:" RESET_COLOR " [";
-	for (size_t i = 0; i < _index.size(); i++)
-	{
-		if (i != 0)
-		{
-			std::cout << ", ";
-		}
-		std::cout << _index[i];
-	}
-	std::cout << ']' << std::endl;
-}
-
-void ConfigLocation::printCgi() const
-{
-	std::cout << "  \t" WHITE_BOLD << "CGI:" RESET_COLOR " [";
-	for (size_t i = 0; i < _cgi.size(); i++)
-	{
-		if (i != 0)
-		{
-			std::cout << ", ";
-		}
-		std::cout << _cgi[i].first << ";" << _cgi[i].second;
-	}
-	std::cout << ']' << std::endl;
-}
-
-
-
-
