+++ src/src_normal/config/resources/default.conf
@@ -4,7 +4,7 @@ server {
     listen 8082;
 
     server_name localhost www.localhost;
-    client_body_size 1m;
+    client_body_size 1g;
 
     error_page 404 /path/404.html;
     location / {
@@ -17,7 +17,7 @@ server {
         root data;
         allowed_methods DELETE;
         autoindex off;
-        index home.html;
+        index home;
         cgi .php /php;
     }
     location /test/ {
